requirements.txt
fastapi==0.95.2
uvicorn[standard]
pillow